SCHEMA-521. Provide options that specify how to decode columns.
This commit adds a serializable class, DecodeConfig, that can be used
to specify how cells from a Kiji table are decoded.
ColumnsDecodeConfig is also added, and associates columns of a Kiji
table with a DecodeConfig.

Eventually, this should become the interface used to specify decode
options when opening a table reader factory. In the short-term, it
will be useful for communicating decode settings to MapReduce tasks.